More default shapes, fix duplicate message bug.
BodyConfigurationFactory is now a singleton since it is needed everywhere and holds no state.
Methods that returned ShapeDimensions for Ball and Paddle objects now return Ball and Paddle objects.
Add a bunch of methods to create ball and paddle objects with default dimensions.
Refactored almost all tests to use these new methods, didn't change the levels that are used in complicated tests.

Fix duplicate removebrick messages. Reason for duplicates: when an explosive destroys another brick that brick also triggers its explosive effect,
so the message to remove the brick is added by the big explosive effect and by the default explosive effect.
Fix is using a Set instead of a List to store brickmessages. This also changes parameter type from List to Collection so it doesn't matter
what specific collection is used.