Move lives from level to player.
Player loses life if the ball that went out of bounds has the same playerIndex as the player.
Player lives are set in game.addConnectingPlayer.
Add method to SessionManager to find player by index.
Projectile playerIndex is used to decide who should lose a life.
Add hasTwoBounds parameter to levelstate. If true, the top wall is a sensor callend "ground".
Any test where a player loses a life now must have a player connected to the game.