Fix startball bug due to cookies not being sent.
The fetch that happens to start the ball did not send cookies, so the session was empty.
The AuthenticationFilter puts a user in the session if no user is in the session, so there was still
a player in the session with playerIndex 1, so there were no problems.
The problem presented itself when the second player in a multiplayer game tried to start a game.
Also made some things serializable and transient to fix session errors.