Add winnername to levelcomplete message.
In breakoutworld.handle ballbrickcontact we set the playerindex to the owner of the ball that hit the brick.
In breakoutEffectHandler for explossiveeffect we copy the playerindex from the centrebrick to all bricks that will be removed.
In level.removeBrick we can then get the playerindex and pass it on so the sessionmanager can find out what player won.